+++ stroom-index-server/src/main/java/stroom/search/server/shard/IndexShardSearchTask.java
@@ -18,18 +18,14 @@ package stroom.search.server.shard;
 
 import org.apache.lucene.search.Query;
 import org.apache.lucene.util.Version;
-import stroom.search.coprocessor.Error;
 import stroom.search.coprocessor.Receiver;
-import stroom.search.coprocessor.Values;
-
-import java.util.function.Consumer;
 
 class IndexShardSearchTask {
     private final IndexShardQueryFactory queryFactory;
     private final long indexShardId;
     private final String[] fieldNames;
     private final Receiver receiver;
-    private final Tracker tracker;
+    private final HitCount hitCount;
     private int shardNumber;
     private int shardTotal;
 
@@ -37,12 +33,12 @@ class IndexShardSearchTask {
                          final long indexShardId,
                          final String[] fieldNames,
                          final Receiver receiver,
-                         final Tracker tracker) {
+                         final HitCount hitCount) {
         this.queryFactory = queryFactory;
         this.indexShardId = indexShardId;
         this.fieldNames = fieldNames;
         this.receiver = receiver;
-        this.tracker = tracker;
+        this.hitCount = hitCount;
     }
 
     IndexShardQueryFactory getQueryFactory() {
@@ -61,8 +57,8 @@ class IndexShardSearchTask {
         return receiver;
     }
 
-    Tracker getTracker() {
-        return tracker;
+    HitCount getHitCount() {
+        return hitCount;
     }
 
     int getShardNumber() {